requirements.txt
dash>=1.0.0
gunicorn>=19.9.0
Flask>=1.0.3
dash_table>=4.0.0
requests>=2.17.1
pandas>=0.23.4
dash-bootstrap-components>=0.6.3
PyYAML >= 5.1.1
yfinance >= 0.1.44
selenium >= 3.141.0
beautifulsoup4 >= 4.7.1
lxml >= 4.4.0